# tp-final/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulso API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/proyectos": {
      "get": {
        "tags": [
          "proyectos"
        ],
        "summary": "Index",
        "operationId": "index_api_proyectos_get",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Estado"
            }
          },
          {
            "name": "responsable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Responsable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProyectoResumen"
                  },
                  "title": "Response Index Api Proyectos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "proyectos"
        ],
        "summary": "Create",
        "operationId": "create_api_proyectos_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProyectoIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProyectoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proyectos/{identifier}": {
      "get": {
        "tags": [
          "proyectos"
        ],
        "summary": "Detail",
        "operationId": "detail_api_proyectos__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProyectoDetalle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "proyectos"
        ],
        "summary": "Update",
        "operationId": "update_api_proyectos__identifier__put",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProyectoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProyectoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "proyectos"
        ],
        "summary": "Remove",
        "operationId": "remove_api_proyectos__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proyectos/{identifier}/tareas": {
      "get": {
        "tags": [
          "tareas"
        ],
        "summary": "Index",
        "operationId": "index_api_proyectos__identifier__tareas_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TareaOut"
                  },
                  "title": "Response Index Api Proyectos  Identifier  Tareas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tareas"
        ],
        "summary": "Create",
        "operationId": "create_api_proyectos__identifier__tareas_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TareaIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TareaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tareas/{identifier}": {
      "get": {
        "tags": [
          "tareas"
        ],
        "summary": "Detail",
        "operationId": "detail_api_tareas__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TareaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tareas"
        ],
        "summary": "Update",
        "operationId": "update_api_tareas__identifier__put",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TareaIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TareaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tareas"
        ],
        "summary": "Remove",
        "operationId": "remove_api_tareas__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsumoListado": {
        "properties": {
          "consumo_id": {
            "type": "integer",
            "title": "Consumo Id"
          },
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "recurso_id": {
            "type": "integer",
            "title": "Recurso Id"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "horas_consumidas": {
            "type": "number",
            "title": "Horas Consumidas"
          },
          "tarea": {
            "type": "string",
            "title": "Tarea"
          },
          "rol_id": {
            "type": "integer",
            "title": "Rol Id"
          },
          "proyecto_nombre": {
            "type": "string",
            "title": "Proyecto Nombre"
          },
          "recurso_nombre": {
            "type": "string",
            "title": "Recurso Nombre"
          },
          "rol_descripcion": {
            "type": "string",
            "title": "Rol Descripcion"
          }
        },
        "type": "object",
        "required": [
          "consumo_id",
          "proyecto_id",
          "recurso_id",
          "fecha_inicio",
          "fecha_fin",
          "horas_consumidas",
          "tarea",
          "rol_id",
          "proyecto_nombre",
          "recurso_nombre",
          "rol_descripcion"
        ],
        "title": "ConsumoListado"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProyectoDetalle": {
        "properties": {
          "proyecto": {
            "$ref": "#/components/schemas/ProyectoResumen"
          },
          "consumos": {
            "items": {
              "$ref": "#/components/schemas/ConsumoListado"
            },
            "type": "array",
            "title": "Consumos"
          },
          "por_recurso": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Por Recurso"
          },
          "por_rol": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Por Rol"
          }
        },
        "type": "object",
        "required": [
          "proyecto",
          "consumos",
          "por_recurso",
          "por_rol"
        ],
        "title": "ProyectoDetalle"
      },
      "ProyectoIn": {
        "properties": {
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "proyecto_nombre": {
            "type": "string",
            "title": "Proyecto Nombre"
          },
          "horas_requeridas": {
            "type": "number",
            "title": "Horas Requeridas"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "proyect_status": {
            "type": "string",
            "title": "Proyect Status"
          },
          "porcentaje_avance": {
            "type": "number",
            "title": "Porcentaje Avance"
          }
        },
        "type": "object",
        "required": [
          "fecha_inicio",
          "fecha_fin",
          "proyecto_nombre",
          "horas_requeridas",
          "proyect_status",
          "porcentaje_avance"
        ],
        "title": "ProyectoIn"
      },
      "ProyectoOut": {
        "properties": {
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "proyecto_nombre": {
            "type": "string",
            "title": "Proyecto Nombre"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "horas_requeridas": {
            "type": "number",
            "title": "Horas Requeridas"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "proyect_status": {
            "type": "string",
            "title": "Proyect Status"
          },
          "porcentaje_avance": {
            "type": "number",
            "title": "Porcentaje Avance"
          }
        },
        "type": "object",
        "required": [
          "proyecto_id",
          "proyecto_nombre",
          "fecha_inicio",
          "fecha_fin",
          "horas_requeridas",
          "owner_id",
          "proyect_status",
          "porcentaje_avance"
        ],
        "title": "ProyectoOut"
      },
      "ProyectoResumen": {
        "properties": {
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "proyecto_nombre": {
            "type": "string",
            "title": "Proyecto Nombre"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "horas_requeridas": {
            "type": "number",
            "title": "Horas Requeridas"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "proyect_status": {
            "type": "string",
            "title": "Proyect Status"
          },
          "porcentaje_avance": {
            "type": "number",
            "title": "Porcentaje Avance"
          },
          "responsable": {
            "type": "string",
            "title": "Responsable"
          },
          "horas_consumidas": {
            "type": "number",
            "title": "Horas Consumidas"
          },
          "saldo": {
            "type": "number",
            "title": "Saldo"
          },
          "exceso": {
            "type": "number",
            "title": "Exceso"
          },
          "porcentaje_consumo": {
            "type": "number",
            "title": "Porcentaje Consumo"
          }
        },
        "type": "object",
        "required": [
          "proyecto_id",
          "proyecto_nombre",
          "fecha_inicio",
          "fecha_fin",
          "horas_requeridas",
          "owner_id",
          "proyect_status",
          "porcentaje_avance",
          "responsable",
          "horas_consumidas",
          "saldo",
          "exceso",
          "porcentaje_consumo"
        ],
        "title": "ProyectoResumen"
      },
...
      "TareaIn": {
        "properties": {
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "tarea_nombre": {
            "type": "string",
            "title": "Tarea Nombre"
          },
          "porcentaje_avance": {
            "type": "number",
            "title": "Porcentaje Avance"
          },
          "recurso_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Id"
          }
        },
        "type": "object",
        "required": [
          "fecha_inicio",
          "fecha_fin",
          "tarea_nombre",
          "porcentaje_avance"
        ],
        "title": "TareaIn"
      },
      "TareaOut": {
        "properties": {
          "tarea_id": {
            "type": "integer",
            "title": "Tarea Id"
          },
          "proyecto_id": {
            "type": "integer",
            "title": "Proyecto Id"
          },
          "tarea_nombre": {
            "type": "string",
            "title": "Tarea Nombre"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "porcentaje_avance": {
            "type": "number",
            "title": "Porcentaje Avance"
          },
          "recurso_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Id"
          },
          "recurso_nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Nombre"
          }
        },
        "type": "object",
        "required": [
          "tarea_id",
          "proyecto_id",
          "tarea_nombre",
          "fecha_inicio",
          "fecha_fin",
          "porcentaje_avance",
          "recurso_id",
          "recurso_nombre"
        ],
        "title": "TareaOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}